# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Creates a new user account and sends an email confirmation link.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/confirm-email": {
      "get": {
        "description": "Validates the confirmation token and marks the user's email as confirmed.",
        "operationId": "AuthController_confirmEmail",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Email confirmed successfully"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired confirmation token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Confirm email address",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/resend-confirmation": {
      "post": {
        "description": "Resends the email confirmation link. No-op if the email is already confirmed or does not exist.",
        "operationId": "AuthController_resendConfirmation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Confirmation email sent (or silently skipped if already confirmed)"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Resend confirmation email",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates a confirmed user and returns an access + refresh token pair.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Email not confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Log in",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Issues a new access + refresh token pair using a valid refresh token.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, or reused refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Refresh tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "Sends a password reset email. No-op if the email is not registered.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Reset email sent (or silently skipped if email not found)"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Request password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "description": "Resets the user password using a valid reset token.",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password reset successfully"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired reset token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Reset password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Revokes all active refresh tokens for the authenticated user.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Log out",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Returns the authenticated user's profile from the JWT payload.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user payload",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "sub": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "auth"
        ]
      }
    },
    "/videos/init": {
      "post": {
        "description": "Creates a new video draft record with a unique slug and issues direct S3/MinIO upload URLs (single PUT or multipart upload parts).",
        "operationId": "VideosController_initUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Video draft created and presigned upload URL(s) issued."
          },
          "400": {
            "description": "Validation error or file size exceeds maximum 10GB."
          },
          "401": {
            "description": "Unauthorized — requires valid JWT access token."
          },
          "404": {
            "description": "Channel not found for current user."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initialize video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/complete": {
      "post": {
        "description": "Validates video ownership and state, transitions status to PROCESSING, and enqueues a background transcode/thumbnail job.",
        "operationId": "VideosController_completeUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID of the video record",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload confirmed and background processing job queued."
          },
          "400": {
            "description": "Video is not in DRAFT/UPLOADING state."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden — user does not own this channel."
          },
          "404": {
            "description": "Video not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Complete video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{slug}/stream": {
      "get": {
        "description": "Streams video directly from storage. Supports HTTP Range header for seeking with 206 Partial Content response.",
        "operationId": "VideosController_getStream",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Unique public slug of the video",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Range",
            "in": "header",
            "description": "HTTP Byte range header (e.g. bytes=0-1048576)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full video stream."
          },
          "206": {
            "description": "Partial content byte stream according to requested Range."
          },
          "400": {
            "description": "Video is not in READY state."
          },
          "404": {
            "description": "Video not found."
          }
        },
        "summary": "Stream video content",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{slug}/download": {
      "get": {
        "description": "Provides direct file download of the video with Content-Disposition attachment header.",
        "operationId": "VideosController_getDownload",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Unique public slug of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary video file stream."
          },
          "400": {
            "description": "Video is not in READY state."
          },
          "404": {
            "description": "Video not found."
          }
        },
        "summary": "Download original video file",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{slug}": {
      "get": {
        "description": "Returns public video details and status by its unique slug.",
        "operationId": "VideosController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Unique public slug of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Video metadata and status."
          },
          "404": {
            "description": "Video not found."
          }
        },
        "summary": "Get video details by slug",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "ResendConfirmationDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {}
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "InitUploadDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the video",
            "maxLength": 255,
            "example": "My First Video"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the video",
            "example": "This is a description of my first video."
          },
          "file_name": {
            "type": "string",
            "description": "Original name of the video file",
            "example": "gameplay.mp4"
          },
          "file_size": {
            "type": "number",
            "description": "Size of the file in bytes (up to 10GB = 10,737,418,240 bytes)",
            "minimum": 1,
            "maximum": 10737418240,
            "example": 52428800
          },
          "content_type": {
            "type": "string",
            "description": "MIME content type of the video file",
            "example": "video/mp4"
          }
        },
        "required": [
          "title",
          "file_name",
          "file_size",
          "content_type"
        ]
      },
      "MultipartPartDto": {
        "type": "object",
        "properties": {
          "PartNumber": {
            "type": "number",
            "description": "Part number in the multipart upload sequence",
            "example": 1
          },
          "ETag": {
            "type": "string",
            "description": "ETag returned by S3 upon uploading the part",
            "example": "\"d41d8cd98f00b204e9800998ecf8427e\""
          }
        },
        "required": [
          "PartNumber",
          "ETag"
        ]
      },
      "CompleteUploadDto": {
        "type": "object",
        "properties": {
          "parts": {
            "description": "Array of uploaded parts with their ETags for multipart uploads",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MultipartPartDto"
            }
          }
        }
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}